fix: refactor reservations overview page, split tabs into separate files (wip)
